refactor, sorta - sort ascending
